still working, moved index logic over to product page
