+++ BangazonSite/Controllers/HomeController.cs
@@ -3,32 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using BangazonSite.Data;
 
 namespace BangazonSite.Controllers
 {
     public class HomeController : Controller
     {
-        private readonly ApplicationDbContext _context;
-
-        public HomeController (ApplicationDbContext context)
+        public IActionResult Index()
         {
-            _context = context;
-        }
-
-        //public IActionResult Index()
-        public ViewResult Index(string searchString)
-        {
-            var products = from p in _context.Product
-                           select p;
-
-            if (!String.IsNullOrEmpty(searchString))
-            {
-                products = products.Where(p => p.Title.Contains(searchString)
-                                       || p.Description.Contains(searchString));
-            }
-
-            return View(products.ToList());
+            return View();
         }
 
         public IActionResult About()
